db(schema): update the createdAt field for the User's model
update the createdAt field to be a required field and to remove the sql
default metadata.

This fixes an issue with planetscale where creating new db branches off
the main branch takes longer than expected (more than 23 hours for me).
I got the idea for this fix from the planetscale support team.